Feat/auth (#46)
* feat: New nav, account page

* feat: Authentication & plenty of other things

+++ src/routes/api/list/item/+server.ts
@@ -4,7 +4,7 @@ import type { RequestHandler } from './$types';
 import { ListZ, RestaurantZ } from '$lib/types';
 import z from 'zod';
 
-export const POST: RequestHandler = async ({ request }) => {
+export const POST: RequestHandler = async ({ request, locals }) => {
 	const body = await request.json();
 	const schema = z.object({
 		restaurantId: RestaurantZ.shape.id,
@@ -15,12 +15,17 @@ export const POST: RequestHandler = async ({ request }) => {
 		return json({ error: 'Invalid request body', details: parseResult.error }, { status: 400 });
 	}
 	const { restaurantId, listId } = parseResult.data;
+	const user = locals.user!;
+	const isOwner = await ListDAO.checkOwnership(listId, user.id);
+	if (!isOwner) {
+		return json({ error: 'Unauthorized' }, { status: 403 });
+	}
 	await ListDAO.addRestaurantToList(restaurantId, listId);
 
 	return json({ success: true }, { status: 201 });
 };
 
-export const DELETE: RequestHandler = async ({ request }) => {
+export const DELETE: RequestHandler = async ({ request, locals }) => {
 	const body = await request.json();
 	const schema = z.object({
 		restaurantId: RestaurantZ.shape.id,
@@ -31,6 +36,11 @@ export const DELETE: RequestHandler = async ({ request }) => {
 		return json({ error: 'Invalid request body', details: parseResult.error }, { status: 400 });
 	}
 	const { restaurantId, listId } = parseResult.data;
+	const user = locals.user!;
+	const isOwner = await ListDAO.checkOwnership(listId, user.id);
+	if (!isOwner) {
+		return json({ error: 'Unauthorized' }, { status: 403 });
+	}
 	await ListDAO.removeRestaurantFromList(restaurantId, listId);
 	return json({ success: true }, { status: 200 });
 };
